fix ruff errors

+++ scal3/cal_types/gregorian_proleptic.py
@@ -68,7 +68,7 @@ from math import floor
 
 
 def ifloor(x):
-	return int(floor(x))
+	return floor(x)
 
 
 epoch = 1721426
